Create subdirectories for autosave
Saving to the Desktop like Screen Shot sounded great at the time,
and I turn of Desktop icons so it didn't bother me, but really it
was a mess and a bit silly considering the number of autosave
clippings that can be produced. So it is time to create year and
month directories to organize them.

Manually saved clippings don't get year and month directories
because they are an explicit user action and the user can handle
them (and might well not want to have to hunt for them).

Surely some users will want year/month/day directories, and some
only year. Perhaps some none. I can't bring myself to make the
settings panel more complicated, so year/month seemed to strike
a balance.